server_flask/requirements.txt
attrs==24.2.0
bcrypt==4.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
mysql-connector-python
packaging==24.2
pycparser==2.22
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
Werkzeug==3.1.3
zipp==3.21.0
python-dotenv
gunicorn